Dockerized build and deploy process of webapp and Redis
Using Docker and Docker-Compose it is now possible to build the
webapp and deploy it automatically connected with a Redis instance
without having to install Gradle or the project dependencies in
the development machine.

Previosuly, the connection to Redis from Spring was done using
the default values of the framework, that used localhost as IP
of the Redis instance, which turns unusable when putting the
services in different containers. To solve this environmental
variables have been added through the .env file. Note that the
variables in this file will override any Spring variable, be it
default or manually specified in the application.properties file,
when served through docker-compose.

Future optimizations need to be made to the Dockerfile to improve
efficiency of the build process. Since so far I haven't found
a way to precompile the dependencies of the project with the
build.gradle file alone, the image has to build everything at
source file modification, which makes the build process much longer
and may affect the utility of this containerized environment
for a developper.